Adds dist files
